pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cert-agent-exp"
version = "0.1.0"
description = "Certificate-gated defenses for adversarial control-flow injection in LLM agents"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "rich>=13.7",
  "orjson>=3.9",
  "regex>=2024.0",
  "scikit-learn>=1.4",
  "matplotlib>=3.8",
  "sentence-transformers>=3.0",
  "faiss-cpu>=1.8.0",
  "datasets>=2.19",
]

[project.scripts]
cert-agent-exp = "cert_agent_exp.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
pythonpath = ["src"]
